requirements.txt
fastapi
uvicorn
python-dotenv
supabase-py
python-jose
requests
httpx
